[binary] Support returning arbitrary mime types
If the produces field is set to one mime type and the return is binary (type=string,format=binary) then this mime type is used for the returned value.

If the produces field has more than one value the the return type of the abstract method becomes MimeWrapper. The context has a method wrap(mime,byte[]) to create a MimeWrapper.

If the mime type is application/json and the return type is binary then the byte[] is assumed to hold json.