[cors] Fixed
https://github.com/ennexos/local-device-sdk/issues/17

I've traced it down to the CORSImplementation.java, l230, where it says

If each of the header field-names is a simple header and none is
Content-Type, this step may be skipped.

However, the following if statement does exactly the opposite: It only adds the allow-headers header when not all of the fields are simple headers and none of them is content-type.

This makes it impossible to develop the UI-sdk against the live LDx-environment. Is there a simple way around this, apart from completely disabling CORS in my browser (chrome --disable-web-security)?