[gen] Paths with overlapping parameters
If a path had overlapping parameters then the dispatch
was not valid. The code did not consume all parameters
after it detected 1 parameter and then dispatches on the
arity of the parameters.

The fix was to consume all parameters once one is detected and 
store them on arity. Appropriate exceptions are thrown when
overlapping paths conflict.