[validate] Updated handling of require/validate
At code generation time it is possible to not
generate the require & validate calls. For this,
there is now an Options set in the Configuration.

The Options is an enum with NOREQUIREMENT, NOVALIDATION
Clearly this enum can be extended in the future.

Second, there are two configuration properties
for config admin:

- checkRequirements
- checkValidate

These properties are set in the OpenAPIRuntime
into fields:

- validate
- require

Additionally, there are 2 protected boolean fields
in OpenAPIBase, the class that implementations inherit.

- validate
- require

The OpenAPIContext.require(...) now will check
if the OpenAPIRuntime.require && OpenAPIBase.require
are set to true. This only controls the reporting,
the returned boolean indicates the status if the
value is present or not. (not null).

The OpenAPIContext.validate now will check the 
OpenAPIRuntime.validate && OpenAPIBase.validate. Only
when both are true, the validation is executed. If no
validation is executed, true is returned otherwise
the status of the validation