[validators/x-type] X-type got called on a validator
The code would call the validator on an x-type but
this was of course not there, causing compile errors.

Test code from https://github.com/kodonnell-heila. (Thanks!)

Added a test case in the generator because it was
easier to debug.

Turns out that the hasValidator() for Object was
not checking for the x-type. Now, if there is an
x-type set (actualType) then we will not validate.